requirements.txt
cassandra-driver==3.25.0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
export==0.1.2
Flask==2.0.1
Flask-WTF==0.15.1
geomet==0.2.1.post1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
Werkzeug==2.0.1
wincertstore==0.2
wtf==0.1
WTForms==2.3.3